+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
   compatibility with short-cut keys and password fields.
 - tip: re-enable support for completely disable input processing via Ctrl+Space.
 - tip: initial support for using Ctrl+Delete to unlearn phrases.
+- tip: support using custom "text" keybinding to insert any text.
 - tip: enable chewing.auto_snapshot_selections by default.
 - tip: support zh-CN language profile.
 - prefs: preferences now only show fonts that can render Chinese characters.